[FIX] no stateReason for empty hash
It has happened for me that describe_instances received empty hashes for
a stateReason, i.e. both code and message where undef. In this case an
exception is thrown because undef violates the type constraints for the
StateReason objects. If an empty hash is given, we now act like no
stateReason was given at all.